backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
filelock==3.10.7
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
Pillow==9.5.0
platformdirs==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
simplejson==3.18.4
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
